% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_microbe_pathway_network.R
\name{construct_microbe_pathway_network_internal}
\alias{construct_microbe_pathway_network_internal}
\title{Internal helper for Microbe-pathway network construction}
\usage{
construct_microbe_pathway_network_internal(
  contrib_file,
  metadata_file,
  taxonomy_file,
  output_file,
  file_type = c("csv", "tsv"),
  filtering = c("unfiltered", "mean", "median", "top10\%", "top25\%", "top50\%",
    "top75\%")
)
}
\arguments{
\item{contrib_file}{A character string specifying the path to the contribution data file.}

\item{metadata_file}{A character string specifying the path to the sample metadata file.}

\item{taxonomy_file}{A character string specifying the path to the taxonomy data file.}

\item{output_file}{A character string specifying the path to the directory
where the output CSV files will be saved.}

\item{file_type}{A character string indicating the type of input files.}

\item{filtering}{A character string specifying the filtering method to apply.}
}
\value{
A character vector of paths to the generated microbe-pathway network CSV files.
}
\description{
This function processes microbial contribution, metadata, and taxonomy data
to construct a microbe-pathway network, calculating relative contributions
of taxa to specific functions. It supports various filtering options
to focus on the most significant contributions. This is an internal function
primarily used by `construct_multi_layered_network_full` and wrapped by
`construct_microbe_pathway_network`.
}
\keyword{internal}
