% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determine_node_groups}
\alias{determine_node_groups}
\title{Determine node groups from network data and falls back to regex}
\usage{
determine_node_groups(nodes, network_data, source_col, target_col)
}
\description{
Determine node groups from network data and falls back to regex
}
\keyword{internal}
