% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_microbe_pathway_network.R
\name{.mpn_apply_filter}
\alias{.mpn_apply_filter}
\title{Internal helper: Apply MPN filtering logic}
\usage{
.mpn_apply_filter(res, mpn_filtering)
}
\description{
Internal helper: Apply MPN filtering logic
}
\keyword{internal}
