% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ctrl_drag_js}
\alias{get_ctrl_drag_js}
\title{Get JavaScript for Ctrl-drag Multi-Node Selection}
\usage{
get_ctrl_drag_js()
}
\description{
Get JavaScript for Ctrl-drag Multi-Node Selection
}
\keyword{internal}
