% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_identification.R
\name{iden_hub}
\alias{iden_hub}
\title{Hub Identification using Maximal Clique Centrality (MCC) algorithm}
\usage{
iden_hub(
  multi_layered_network_file,
  output_directory,
  top_n_hubs = 5,
  visualize = TRUE
)
}
\arguments{
\item{multi_layered_network_file}{Path to the multi-layered network CSV/TSV
(output from `con_mln`). Must contain 'from'/'to' or 'Feature1'/'Feature2' columns.}

\item{output_directory}{Path to save results.}

\item{top_n_hubs}{Integer. Number of top hub nodes to highlight in the HTML visualization.}

\item{visualize}{Logical. If TRUE, generates interactive HTML.}
}
\value{
Invisible NULL. Results are written to `output_directory`.
}
\description{
Hub Identification using Maximal Clique Centrality (MCC) algorithm
}
\details{
Identifies influential hub nodes in a multi-layered network using the
Maximal Clique Centrality (MCC) algorithm. Exports a ranked CSV of all
node scores and optionally generates an interactive HTML network colored
by MCC score with a customizable palette.
}
