% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_widget_safe}
\alias{save_widget_safe}
\title{Save an htmlwidget safely, falling back to non-selfcontained if pandoc is missing}
\usage{
save_widget_safe(widget, file, title = "NUIMM")
}
\arguments{
\item{widget}{The htmlwidget to save.}

\item{file}{Output file path.}

\item{title}{HTML page title.}
}
\description{
Save an htmlwidget safely, falling back to non-selfcontained if pandoc is missing
}
\keyword{internal}
