% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_multi_layered_network.R
\name{con_mln_int}
\alias{con_mln_int}
\title{Internal MLN Integration & Visualization Helper}
\usage{
con_mln_int(gsea_file, mpn_file, ppn_file, pmn_file, output_dir)
}
\arguments{
\item{gsea_file}{Character path to GSEA file.}

\item{mpn_file}{Character path to MPN file.}

\item{ppn_file}{Character path to PPN file.}

\item{pmn_file}{Character path to PMN file.}

\item{output_dir}{Character path to output directory.}
}
\value{
Final network file path.
}
\description{
Merges MPN, PPN, and PMN layers and visualizes.
}
\keyword{internal}
