% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_microbe_pathway_network.R
\name{con_mpn_int}
\alias{con_mpn_int}
\title{Internal Microbe-Pathway Network Helper}
\usage{
con_mpn_int(
  path_con_file,
  metadata_file,
  taxonomy_file,
  output_dir,
  mpn_filtering = "top10\%"
)
}
\arguments{
\item{path_con_file}{Character path to processed contribution file.}

\item{metadata_file}{Character path to metadata file.}

\item{taxonomy_file}{Character path to taxonomy file.}

\item{output_dir}{Character path to output directory.}

\item{mpn_filtering}{Filtering options: "unfiltered", "mean", "median", "top10%", "top25%", "top50%", "top75%".}
}
\value{
Vector of MPN file paths.
}
\description{
Calculates relative contribution of taxa to functions.
}
\keyword{internal}
