% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pathway_metabolite_network.R
\name{con_pmn_int}
\alias{con_pmn_int}
\title{Internal Pathway-Metabolite Network Helper}
\usage{
con_pmn_int(
  path_abun_file,
  met_con_file,
  gsea_file,
  metadata_file,
  output_dir,
  pmn_corr_method = c("spearman", "pearson", "kendall"),
  pmn_filter_by = "p_value",
  pmn_corr_cutoff = 0.3,
  pmn_pvalue_cutoff = 0.05,
  pmn_q_value_cutoff = 0.05,
  pmn_p_adjust_method = "fdr"
)
}
\description{
Internal Pathway-Metabolite Network Helper
}
\keyword{internal}
