% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pathway_metabolite_network.R
\name{con_pmn_int}
\alias{con_pmn_int}
\title{Internal Pathway-Metabolite Network Helper}
\usage{
con_pmn_int(
  path_abun_file,
  met_con_file,
  gsea_file,
  metadata_file,
  output_dir,
  corr_method = c("spearman", "pearson", "kendall"),
  pmn_filter_by = "p_value",
  corr_cutoff = 0.3,
  pvalue_cutoff = 0.05,
  q_value_cutoff = 0.05,
  p_adjust_method = "fdr"
)
}
\arguments{
\item{path_abun_file}{Character path to pathway abundance file.}

\item{met_con_file}{Character path to metabolite concentration file.}

\item{gsea_file}{Character path to GSEA file.}

\item{metadata_file}{Character path to metadata file.}

\item{output_dir}{Character path to output directory.}

\item{corr_method}{Correlation options: "spearman", "pearson", "kendall".}

\item{pmn_filter_by}{Filter options: "none", "p_value", "q_value".}

\item{corr_cutoff}{Numeric correlation cutoff.}

\item{pvalue_cutoff}{Numeric p-value cutoff.}

\item{q_value_cutoff}{Numeric q-value cutoff.}

\item{p_adjust_method}{P-adj options: "fdr", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "none".}
}
\value{
Vector of PMN file paths.
}
\description{
Correlates pathway abundance and metabolites.
}
\keyword{internal}
