% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pathway_pathway_network.R
\name{con_ppn_int}
\alias{con_ppn_int}
\title{Internal Pathway-Pathway Network Helper}
\usage{
con_ppn_int(
  gene_abun_file,
  metadata_file,
  map_file,
  output_dir,
  da_method = c("deseq2", "edger", "maaslin2", "simple"),
  map_database = c("kegg", "metacyc"),
  rank_by = c("signed_log_pvalue", "log2foldchange", "pvalue"),
  p_adjust_method = "fdr",
  pvalue_cutoff = 0.05,
  jaccard_cutoff = 0.2,
  min_gs_size = 10,
  max_gs_size = 500,
  eps = 1e-10
)
}
\arguments{
\item{gene_abun_file}{Character path to gene abundance file.}

\item{metadata_file}{Character path to metadata file.}

\item{map_file}{Character path to mapping file.}

\item{output_dir}{Character path to output directory.}

\item{da_method}{DA method options: "deseq2", "edger", "maaslin2", "simple".}

\item{map_database}{Database options: "kegg", "metacyc".}

\item{rank_by}{GSEA ranking options: "signed_log_pvalue", "log2foldchange", "pvalue".}

\item{p_adjust_method}{GSEA p-adj options: "fdr", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "none".}

\item{pvalue_cutoff}{Numeric p-value cutoff.}

\item{jaccard_cutoff}{Numeric Jaccard cutoff.}

\item{min_gs_size}{Numeric min gene set size.}

\item{max_gs_size}{Numeric max gene set size.}

\item{eps}{Numeric GSEA epsilon.}
}
\value{
List of GSEA and Jaccard file paths.
}
\description{
Runs DA, GSEA, and Jaccard index calculation.
}
\keyword{internal}
