% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pathway_pathway_network.R
\name{con_ppn_int}
\alias{con_ppn_int}
\title{Internal helper for Pathway-pathway network construction}
\usage{
con_ppn_int(
  abundance_file,
  metadata_file,
  map_file,
  output_file,
  file_type = c("csv", "tsv"),
  pvalueCutoff,
  pAdjustMethod = c("fdr", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
    "none"),
  rank_by = c("signed_log_pvalue", "log2FoldChange")
)
}
\arguments{
\item{abundance_file}{A character string specifying the path to the
gene abundance data file.}

\item{metadata_file}{A character string specifying the path to the sample metadata file.}

\item{map_file}{A character string specifying the path to the
pathway-to-gene mapping file.}

\item{output_file}{A character string specifying the path to the directory
where output CSV files (GSEA results and Jaccard indices) will be saved.}

\item{file_type}{A character string indicating the type of input files.}

\item{pvalueCutoff}{A numeric value specifying the adjusted p-value cutoff
for determining significance in GSEA results.}

\item{pAdjustMethod}{A character string specifying the method for p-value
adjustment.}

\item{rank_by}{A character string specifying the method to rank genes for GSEA.}
}
\value{
A list containing:
  \describe{
    \item{gsea_paths}{A character vector of paths to the generated GSEA results CSV files.}
    \item{jaccard_paths}{A character vector of paths to the generated Pathway-Pathway Jaccard index CSV files.}
  }
}
\description{
This function performs differential expression analysis (using DESeq2) and
Gene Set Enrichment Analysis (GSEA) to identify enriched pathways and
subsequently computes Jaccard indices to quantify overlap between core
enrichment genes of significant pathways. This is an internal function
primarily used by `con_mln_all` and wrapped by
`con_ppn`.
}
\keyword{internal}
