% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_pathway_pathway_network.R
\name{con_ppn_int}
\alias{con_ppn_int}
\title{Internal Pathway-Pathway Network Helper}
\usage{
con_ppn_int(
  gene_abun_file,
  metadata_file,
  map_file,
  output_dir,
  ppn_da_method = c("deseq2", "edger", "maaslin2", "simple"),
  ppn_map_database = c("kegg", "metacyc", "custom"),
  ppn_rank_by = c("signed_log_pvalue", "log2foldchange", "pvalue"),
  ppn_p_adjust_method = "fdr",
  ppn_pvalue_cutoff = 0.05,
  ppn_jaccard_cutoff = 0.2
)
}
\arguments{
\item{gene_abun_file}{Character path to gene abundance file.}

\item{metadata_file}{Character path to metadata file.}

\item{map_file}{Character path to mapping file.}

\item{output_dir}{Character path to output directory.}

\item{ppn_da_method}{DA method options: "deseq2", "edger", "maaslin2", "simple".}

\item{ppn_map_database}{Database options: "kegg", "metacyc", "custom".}

\item{ppn_rank_by}{GSEA ranking options: "signed_log_pvalue", "log2foldchange", "pvalue".}

\item{ppn_p_adjust_method}{GSEA p-adj options: "fdr", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "none".}

\item{ppn_pvalue_cutoff}{Numeric p-value cutoff.}

\item{ppn_jaccard_cutoff}{Numeric Jaccard cutoff.}
}
\value{
List of GSEA and Jaccard file paths.
}
\description{
Runs DA, GSEA, and Jaccard index calculation.
}
\keyword{internal}
