% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinding.R
\name{find_path}
\alias{find_path}
\title{Pathfinding using Dijkstra's algorithm}
\usage{
find_path(
  multi_layered_network_file,
  source_node,
  target_node,
  output_directory,
  file_type = c("csv", "tsv"),
  visualize = TRUE
)
}
\arguments{
\item{multi_layered_network_file}{Path to input CSV/TSV.}

\item{source_node}{Name of the starting node.}

\item{target_node}{Name of the ending node.}

\item{output_directory}{Path to save results.}

\item{file_type}{"csv" or "tsv".}

\item{visualize}{Logical. If TRUE, saves PDF and PNG plots. Defaults to TRUE.}
}
\description{
Pathfinding using Dijkstra's algorithm
}
\details{
This function calculates the shortest path between two nodes using Dijkstra's
algorithm.
}
