% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_identification.R
\name{iden_hub}
\alias{iden_hub}
\title{Hub identification using Maximal Clique Centrality (MCC) algorithm}
\usage{
iden_hub(
  multi_layered_network_file,
  output_directory,
  file_type = c("csv", "tsv"),
  top_n_hubs = NULL,
  visualize = TRUE
)
}
\arguments{
\item{multi_layered_network_file}{Path to input CSV/TSV.}

\item{output_directory}{Path to save results.}

\item{file_type}{"csv" or "tsv".}

\item{top_n_hubs}{Integer. Number of top hubs to filter and visualize.}

\item{visualize}{Logical. If TRUE, saves a PDF plot of the top hubs and their connections. Defaults to FALSE.}
}
\description{
This function identifies hub nodes in a multi-layered network using the
Maximal Clique Centrality (MCC) algorithm and optionally visualizes the results.
}
\details{
[Standard details...]
}
